Rename webpack-plugin to plugin
This avoids the double webpack name. We're going to reuse this for both
server and client.
