+++ packages/react-flight-dom-webpack/package.json
@@ -13,11 +13,11 @@
     "LICENSE",
     "README.md",
     "index.js",
+    "plugin.js",
     "server.js",
     "server.browser.js",
     "server.node.js",
     "server-runtime.js",
-    "webpack-plugin.js",
     "cjs/",
     "umd/"
   ],
